refactor: rename domain modules to template scaffold layout
Re-converge the domain module layout with the fastmcp-server-template
scaffold (matching sibling repos like paperless-mcp), so server.py's
scaffold imports match the template instead of forcing permanent divergence:

- _server_tools.py     -> tools.py
- _server_resources.py -> resources.py
- _server_prompts.py   -> prompts.py
- service.py           -> domain.py   (class ImageService kept — the template's
                                        `Service` is an explicit placeholder you
                                        replace, not a name to conform to)

All import paths (absolute, relative, mock-patch strings) and the pyproject
ruff per-file-ignores updated accordingly. Pure rename — no behaviour change.

Module-layout portion of the server-skeleton re-converge. The
server.py / _server_deps.py refactor (restore the template's
server_lifespan(_mcp) structure; remove the orphaned build_event_store and
_load_server_config shims) follows as a separate, focused PR.

Refs #276. Part of #273.
