+++ scripts/rename.sh
@@ -60,6 +60,6 @@ echo "Done. Next steps:"
 echo "  1. Review changes: git diff"
 echo "  2. Delete this template: rm -rf scripts/"
 echo "  3. Update README.md and TEMPLATE.md with your service details"
-echo "  4. Add your domain logic in src/$PYTHON_MODULE/_server_tools.py"
+echo "  4. Add your domain logic in src/$PYTHON_MODULE/tools.py"
 echo "  5. Update src/$PYTHON_MODULE/_server_deps.py with your service init"
 echo "  6. Run: uv sync && uv run pytest"
