+++ src/image_generation_mcp/domain.py
