+++ tests/test_background.py
@@ -9,11 +9,11 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from image_generation_mcp.domain import ImageService
 from image_generation_mcp.providers.openai import OpenAIImageProvider
 from image_generation_mcp.providers.placeholder import PlaceholderImageProvider
 from image_generation_mcp.providers.sd_webui import SdWebuiImageProvider
 from image_generation_mcp.providers.types import SUPPORTED_BACKGROUNDS
-from image_generation_mcp.service import ImageService
 
 if TYPE_CHECKING:
     from pathlib import Path
